Split successor probe gap sync tests
